Enable ClojureScript REPL and hot reloading of CLJS code
Requiring the shadow-cljs browser REPL components in
metabase.util.devtools is all that was needed to get the REPL working.

But that exposed a problem - the REPL kept disconnecting when edits were
made to CLJS files. The webpack HotModuleReload plugin does not know
how to handle these changes, so it falls back to the safe thing: full
reload.

Therefore this commit adds HMR integration to metabase.lib.js so that
HMR will know that it's safe to simply reload the CLJS code.

Work remaining:
- The attempt to rerender the entire React component tree doesn't seem
  to be working. I'd be glad to know if that's the wrong component, or
  if it somehow only reaches the outer containers and not into the heart
  of the UI.
- Really the HMR integration should be a build-time hook in shadow-cljs
  that does similar integration to each of the `:entries` that are
  exposed to JS. Currently, editing CLJS code outside `src/metabase/lib`
  won't hot reload.